requirements.txt
#---------------------------------------------------------
# py -m pip install -r requirements.txt
#---------------------------------------------------------
ruamel.yaml
fuzzywuzzy[speedup]
ankisync2
anki
pattern
nltk
tqdm
pattern
python-slugify
pandas
numpy
